ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
transformers==4.35.0
torch==2.1.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
python-multipart==0.0.6
tweepy==4.14.0
textblob==0.17.1
langdetect==1.0.9
Pillow==10.1.0